test(trusty-search): isolate the TRUSTY_HNSW_REVIEW_IDLE mutators into their own test binary
`hnsw_idle_demotion_reviews_clean_promoted_store` read the
`TRUSTY_HNSW_REVIEW_IDLE` gate through `demote_vector_store_if_idle` while two
other tests in the SAME lib test binary wrote that process-global variable —
`hnsw_idle_demotion_skips_when_disabled_via_env` and
`hnsw_review_idle_enabled_default_and_env_override`. A writer holding the var
at "0" when the reader reached the gate made the gate return false, demotion
was skipped, and the reader panicked with "expected the idle sweep to demote
the clean, promoted store".

Both mutators move to `tests/hnsw_review_idle_env.rs`, which is its own test
BINARY and therefore its own process, so the 1580-test lib binary now has zero
writers of this variable. Inside the new binary the two are `#[serial]` against
each other and are the only tests present. Both keep their original bodies —
no assertion weakened, no coverage dropped, no `#[ignore]`, no widened
tolerance.

`#[serial]` alone would not have been enough: `setenv` reallocates the C
`environ` array, so a concurrent non-serial `getenv` can still tear. That is
the same conclusion `service::server::list_repo_identity_tests` reached for
#2717, which names one of these very tests as the mutator that broke it.

Proof, with a 300ms sleep inserted while the var reads "0" to widen the race
window (victim test untouched), `cargo test -p trusty-search --lib -- hnsw
--test-threads=16`:
  before: 0/20 passed, 20 failed, all at tests_idle_evict.rs:416
  after:  20/20 passed, 0 failed (same amplifier still in place)

Closes #3769
